task completed
For diabetic customers, Added support for the Diet version of each beverages.
Added support of one another hot drink as Tea.
Modify the Tea item such that it should be a decoratable beverage with a sweetness type, cream and amount of sweetness in spoons count.
Every decoration has a separate additional price added to its base price in the order total.
Modify the Demo to be interactive. Now the system will ask for options in preparing his meal.