fully implemented with instruction
Modify the code such that every concrete shape implements Cloneable along with a support of "equals(Shape a, Shape b)" to compare two shapes and return true if they are same. You can take help from "refactoring.guru" website.
Add a support for a concrete Avatar of Fish or Human to this example which can be cloneable.
Also create some clone of the Fish and Human in the Demo class and compare with each other.